Remove CSS scripting and make elements flexible
Moved most of the CSS scripting from main.js to style.css because
it looks cleaner that way. Not all styling can be moved from
main.js to style.css without affecting the functionality of the page.

The board and buttons were made to be flexible (i.e able to stretch
with the screen) while retaining the width of the cells.

Also changed the border-color of the rainbow colored cells to a more
transparent color for a better color blend.